renamed color.py
